refactor(api): aligning our API with spec
