+++ api/api_test.go
@@ -173,101 +173,6 @@ func TestHealthCheck(t *testing.T) {
 	runHandlerTestCases(t, "health check", HealthCheckHandler, healthCheckCases, true)
 }
 
-func TestServerReadOnlyRoutes(t *testing.T) {
-	if err := confirmQriNotRunning(); err != nil {
-		t.Skip(err.Error())
-	}
-
-	ctx, cancel := context.WithCancel(context.Background())
-	defer cancel()
-
-	prevXformVer := APIVersion
-	APIVersion = "test_version"
-	defer func() {
-		APIVersion = prevXformVer
-	}()
-
-	// bump up log level to keep test output clean
-	golog.SetLogLevel("qriapi", "error")
-	defer golog.SetLogLevel("qriapi", "info")
-
-	// to keep hashes consistent, artificially specify the timestamp by overriding
-	// the dsfs.Timestamp func
-	prev := dsfs.Timestamp
-	defer func() { dsfs.Timestamp = prev }()
-	dsfs.Timestamp = func() time.Time { return time.Date(2001, 01, 01, 01, 01, 01, 01, time.UTC) }
-
-	client := &http.Client{}
-
-	r, err := test.NewTestRepo()
-	if err != nil {
-		t.Fatalf("error allocating test repo: %s", err.Error())
-	}
-
-	// Cannot use TestRunner because we need to set cfg.API.ReadOnly.
-	// TODO(dlong): Add a testRunner call trace that does this correctly.
-	cfg := testcfg.DefaultConfigForTesting()
-	cfg.API.ReadOnly = true
-	defer func() {
-		cfg.API.ReadOnly = false
-	}()
-
-	node, err := p2p.NewQriNode(r, cfg.P2P, event.NilBus, nil)
-	if err != nil {
-		t.Fatal(err.Error())
-	}
-	// TODO (b5) - hack until tests have better instance-generation primitives
-	inst := lib.NewInstanceFromConfigAndNode(ctx, cfg, node)
-	s := New(inst)
-
-	server := httptest.NewServer(NewServerRoutes(s))
-
-	cases := []struct {
-		method    string
-		endpoint  string
-		resStatus int
-	}{
-		// forbidden endpoints
-		{"GET", "/ipfs/", 403},
-		{"GET", "/ipns/", 404},
-		{"POST", "/save", 403},
-		{"PUT", "/save", 403},
-		{"POST", "/remove", 403},
-		{"DELETE", "/remove", 403},
-		{"POST", "/rename", 403},
-		{"PUT", "/rename", 403},
-
-		// active endpoints:
-		{"GET", "/health", 200},
-		{"GET", "/list/peer", 200},
-		// Cannot test connect endpoint until we have peers in this test suite
-		// {"GET", "/connect/QmZePf5LeXow3RW5U1AgEiNbW46YnRGhZ7HPvm1UmPFPwt", 200},
-		// Cannot test endpoint until we have peers in this test suite
-		// {"GET", "/peer", 200},
-		{"GET", "/get/peer/movies", 200},
-	}
-
-	for i, c := range cases {
-
-		req, err := http.NewRequest(c.method, server.URL+c.endpoint, nil)
-		if err != nil {
-			t.Errorf("case %d error creating request: %s", i, err.Error())
-			continue
-		}
-
-		res, err := client.Do(req)
-		if err != nil {
-			t.Errorf("case %d error performing request: %s", i, err.Error())
-			continue
-		}
-
-		if res.StatusCode != c.resStatus {
-			t.Errorf("case %d: %s - %s status code mismatch. expected: %d, got: %d", i, c.method, c.endpoint, c.resStatus, res.StatusCode)
-			continue
-		}
-	}
-}
-
 type handlerMimeMultipartTestCase struct {
 	method    string
 	endpoint  string
