+++ cmd/qri.go
@@ -68,7 +68,6 @@ https://github.com/qri-io/qri/issues`,
 		NewSaveCommand(opt, ioStreams),
 		NewSearchCommand(opt, ioStreams),
 		NewSetupCommand(opt, ioStreams),
-		NewStatsCommand(opt, ioStreams),
 		NewStatusCommand(opt, ioStreams),
 		NewSQLCommand(opt, ioStreams),
 		NewUseCommand(opt, ioStreams),
